requirements.txt
fastapi[standard]
sqlalchemy>=2.0
asyncpg
greenlet
uvicorn[standard]
python-dotenv>=1.0.0
